+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- [ComponentClient.get_executor][yuyo.components.ComponentClient.get_executor]
+  for getting the executor registered globally for a specific custom ID.
+
 ### Changed
 - The ephemeral default is now ignored for message update create initial
   response and defer calls. (This likely won't effect real world behaviour).